requirements.txt
accelerate==0.27.2
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.3.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==22.2.0
audioread==3.0.1
awscli==1.32.63
blinker==1.7.0
boto3==1.34.47
botocore==1.34.63
cachetools==5.3.3
cbor2==5.6.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.4
contourpy==1.2.0
cycler==0.12.1
decorator==4.4.2
deepl==1.17.0
diffusion-core==0.0.65
distro==1.9.0
docutils==0.16
einops==0.7.0
elevenlabs==1.0.0b1
encodec==0.1.1
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.109.2
filelock==3.13.1
fonttools==4.49.0
frozenlist==1.4.1
fsspec==2024.2.0
funcy==2.0
gitdb==4.0.11
GitPython==3.1.42
grpcio==1.53.0
grpcio-tools==1.53.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
huggingface-hub==0.20.3
idna==3.6
imageio==2.34.0
imageio-ffmpeg==0.4.9
importlib-metadata==7.0.1
ipython==8.22.1
jedi==0.19.1
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lazy_loader==0.3
librosa==0.10.1
llvmlite==0.42.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.5
mysql-connector-python==8.3.0
networkx==3.2.1
noisereduce==3.0.0
numba==0.59.0
numpy==1.26.4
openai==1.12.0
opencv-python==4.9.0.80
packaging==23.2
pandas==2.2.1
param==2.0.2
parso==0.8.3
pexpect==4.9.0
pillow==9.5.0
platformdirs==4.2.0
pooch==1.8.1
proglog==0.1.10
prompt-toolkit==3.0.43
protobuf==4.21.12
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.0
pyasn1==0.5.1
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
pydeck==0.8.1b0
pydub==0.25.1
pyfcm==1.5.4
pygame==2.5.2
Pygments==2.17.2
pyparsing==3.1.1
pyproject==1.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
regex==2023.12.25
rembg==2.0.56
requests==2.31.0
rich==13.7.0
rpds-py==0.18.0
rsa==4.7.2
s3transfer==0.10.1
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.12.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soundfile==0.12.1
soxr==0.3.7
stability-sdk==0.8.5
stack-data==0.6.3
starlette==0.36.3
streamlit==1.31.1
stringcase==1.2.0
structlog==21.5.0
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.3.0
tokenizers==0.15.2
toml==0.10.2
toolz==0.12.1
torch==2.2.0
torchaudio==2.2.0
torchvision==0.17.0
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
typing_extensions==4.9.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
uvicorn==0.27.1
validators==0.22.0
watchdog==4.0.0
wcwidth==0.2.13
websockets==12.0
yarl==1.9.4
zipp==3.17.0
